extract the cohortMetadata subcomponent and import in place
